requirements/standard.txt
Django==1.6.5
requests==2.3.0
dj-static==0.0.5
django-static-root-finder==0.1.1
ubuntudesign-asset-mapper==0.7
-e bzr+ssh://bazaar.launchpad.net/+branch/django-openid-auth/@110#egg=django-openid-auth
dj-database-url==0.3.0
psycopg2==2.6.1
python-openid==2.2.5
